batman-adv: Replace gw_reselect 64 bit div with shift
The 64-bit gw_factor is divided by BATADV_TQ_LOCAL_WINDOW_SIZE ** 2 * 64.
But the rest of the calculation has nothing to do with the tq window size
end therefore the calculation is just (tmp_gw_factor / (64 ** 3)).

The problem with 64 bit div is that it doesn't work on systems without
native 64 bit div support. It has to be emulated using do_div or div_u64.
The change in f63c54bba31d ("batman-adv: Avoid u32 overflow during gateway
select") only compiled on such systems because the compiler converted the
div to a (tmp_gw_factor >> 18). Making this explicit avoids having build
problems in the future when BATADV_TQ_LOCAL_WINDOW_SIZE is changed in such
a way that (BATADV_TQ_LOCAL_WINDOW_SIZE ** 2 * 64) is not a power of two.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>